+++ src/compendium.js
@@ -2,6 +2,9 @@ import { MODULE_ID } from "./constants.js";
 import Log from "./log.js";
 
 const COMPENDIUM_MODULE_SETTING = "compendium-module"
+const COMPENDIUM_ALLOW_DELETE_SETTING = "compendium-allow-delete"
+const COMPENDIUM_FORCE_UNIQUE = "compendium-force-unique"
+const COMPENDIUM_SKIP_ADD = "compendium-skip-add"
 
 export function registerCompendiumSettings() {
   const choices = {};
@@ -11,13 +14,36 @@ export function registerCompendiumSettings() {
   })
 
   game.settings.register(MODULE_ID, COMPENDIUM_MODULE_SETTING, {
-    name: "Module to update compendium of",
-    hint: "Name of the module.",
+    name: "Module to store compendium information",
     scope: "world",
     config: true,
     type: String,
     choices,
     default: MODULE_ID
+  });
+
+  game.settings.register(MODULE_ID, COMPENDIUM_ALLOW_DELETE_SETTING, {
+    name: "Delete items from compendium",
+    scope: "world",
+    config: true,
+    type: Boolean,
+    default: true
+  });
+
+  game.settings.register(MODULE_ID, COMPENDIUM_FORCE_UNIQUE, {
+    name: "Force unique names in compendium",
+    scope: "world",
+    config: true,
+    type: Boolean,
+    default: true
+  });
+
+  game.settings.register(MODULE_ID, COMPENDIUM_SKIP_ADD, {
+    name: "Skip adding to compendium",
+    scope: "world",
+    config: true,
+    type: Boolean,
+    default: false
   });
 }
 
@@ -25,7 +51,57 @@ export function getCompendiumModule() {
   return game.settings.get(MODULE_ID, COMPENDIUM_MODULE_SETTING)
 }
 
-export function getItemPack() {
-  // TODO: Be able to specify the name of the pack
-  return game.modules.get(getCompendiumModule()).pack.find(p => p.type === "Item");
+export function shouldDeleteMissingItems() {
+  return !game.settings.get(MODULE_ID, COMPENDIUM_ALLOW_DELETE_SETTING) || game.settings.get(MODULE_ID, COMPENDIUM_SKIP_ADD)
+}
+
+export function forceUniqueNames() {
+  return game.settings.get(MODULE_ID, COMPENDIUM_FORCE_UNIQUE)
+}
+
+export function skipAddToCompendium() {
+  return game.settings.get(MODULE_ID, COMPENDIUM_SKIP_ADD);
+}
+
+export function getPackForDocType(docType) {
+  const pack = game.modules.get(getCompendiumModule()).packs.find(p => p.type === docType);
+  return game.packs.get(`${pack.package}.${pack.name}`);
+}
+
+// export function getPackByNameForDocType(docType, name) {
+//   return game.modules.get(getCompendiumModule()).pack.find(p => p.type === docType && p.name === name);
+// }
+
+export async function getDocsForCompendiumType(docType) {
+  if (!skipAddToCompendium()) {
+    Log.info(getPackForDocType(docType))
+    return (await getPackForDocType(docType).getDocuments())
+  }
+
+  // TODO: Support more than just items
+  switch(docType) {
+    case "Item": {
+      return [...game.items.values()];
+    }
+    default: {
+      Log.error("Tried to load non-compendium docs for unsupported doc type:", docType)
+    }
+  }
+}
+
+export async function deleteCompendiumItems(docType, ids) {
+  // TODO: Delete items imported to the game?
+  const c = getPackForDocType(docType);
+
+  await Promise.all(ids.map(id => c.delete(id)))
+}
+
+export async function addCompendiumItems(docType, items) {
+  const c = getPackForDocType(docType);
+  if (!c) return;
+
+  // TODO: support more than just items
+  const created = await Item.createDocuments(items)
+  await Promise.all(created.map(i => c.importDocument(i)));
+  await Item.deleteDocuments(created.map(i => i.id));
 }