+++ src/constants.js
@@ -1,8 +1,8 @@
 export const MODULE_ID = "cdm"
 
 export const FLAGS = Object.freeze({
-  MANAGED: "managed",
-  ID: "id"
+  ID: "id",
+  FILE_PATH: "filePath"
 });
 
 export const TEMPLATES = Object.freeze({