+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Create the initial project.
 - Basics of handling storage written out.
 - Basic Logging utility created.
+- Load file data from directory.
+- First pass at loading data into the compendium.